Convert some ints to uints and add some tests.
The right-shift operator on a negative signed int is implementation defined,
and gcc on x86 appears to do a sign-extending shift.  This means that -1 >> 1 = -1

writeVarint32 goes into an infinite loop and segfaults if you add this to
your test: testNaked<TProto, int64_t>(0x7fffffffffffffffll);

I changed some ints to uints and added some casts to fix this.
(Basically everywhere the java version uses >>> instead of >>)